ta/sims: add operation for panic triggering
This operation is used for the next set of tests that are
checking the resource cleaning in case of a panic event.
Thus, it makes possible to reproduce those situations whenever
TA context is getting panicked and TEE core has to safely
clear all resources (e.g. sessions, memory) linked to that
specific panicked TA instance.

Signed-off-by: Ovidiu Mihalachi <ovidiu_mihalachi@mentor.com>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>