ta/sims: open another TA session in SIMS
This commit adds the possibility to open another TA session from a TA,
scenario that is required in those tests that are checking
the resource releasing mechanism in case of a panic event.
Thus, this would make possible to check if the secure sessions are
properly removing any reference to the panicked context.

Signed-off-by: Ovidiu Mihalachi <ovidiu_mihalachi@mentor.com>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>