regression: fix memory leakage in case 1010
In regression test case 1010 xtest_tee_test_invalid_mem_access2()
allocates shared memory, but doesn't free it. This can affect other test
cases run after this case.

The shared memory is freed when the context is finalized, but that can
be too late since it normally happens when xtest exits.

Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>