Call TEE_CreatePersistentObject() with a valid pointer
According to GP specification, passing a NULL pointer to
TEE_OpenPersistentObject() and TEE_CreatePersistentObject()
should panic the trusted application.
Current code is using a NULL pointer and expects success.
After creation (and opening) of the new object,
it should also be closed: we cannot expect to do anything on it.

Signed-off-by: Cedric Auger <cedric.auger@provenrun.com>
Reviewed-by: Jerome Forissier <jerome@forissier.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>