github: update actions/workflow file
Sync the GitHub workflow file with the recent changes in optee_os as
described in [1] with the difference that we use a more recent version
of the "actions/stale" workflow (v3.0.10). In short, this updates the
version and add "bug" and "enhancement" as the labels where the GitHub
bot shouldn't mark an issue or pull request as stale ticket.

Link: [1] https://github.com/OP-TEE/optee_os/commit/50bbda3dd3

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Acked-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>