xtest: pkcs11: Modify testcase 1011 for changes in C_Logout()
PKCS#11 specification states that :
When C_Logout successfully executes, any of the application’s
handles to private objects should become invalid (even if a user
is later logged back into the token, those handles remain invalid).
In addition, all private session objects from sessions belonging
to the application should also be destroyed.

One of the find sub-test case was expecting that after logging back
in a session , private session object would be visible which was
a wrong assumption. This has been modified.

Another subcase assumed the handle of the token object returned
would be same as that when the token object was created by that
session. This assumption would hold good only during the first
log in into this session. If a user logs out and then logs in
again, the token object handles would change. So this sub-case
has been moved to be tested during the first login to the session.

Signed-off-by: Ruchika Gupta <ruchika.gupta@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>