Update host/openssl/lib/arm/libcrypto.a with -fPIC
Replace the pre-built Arm 32-bit OpenSSL crypto library with the same,
built with -fPIC for position-independent code. Some "hardened"
environments require PIC, and therefore this new version should provide
wider compatibility.

Note that it is always possible to specify a custom OpenSSL library at
build time via OPTEE_OPENSSL_EXPORT, but having something more likely to
work out of the box is nice.

The build steps for the new libcrypto.a are:

 git clone https://github.com/openssl/openssl.git
 cd openssl
 git checkout OpenSSL_1_0_2o
 ./Configure linux-generic32 -mcpu=cortex-a9 no-shared \
             --cross-compile-prefix=arm-linux-gnueabihf- \
            -fPIC
 make build_libcrypto

Suggested-by: Markus S. Wamser <markus.wamser@mixed-mode.de>
Link: https://github.com/OP-TEE/optee_test/issues/507
Signed-off-by: Jerome Forissier <jerome@forissier.org>
Tested-by: Aleksandar Nikolic <aleksandar.nikolic@mixed-mode.de>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>