gp: disable test ce-7b-17 (gp_40178)
Add a patch to exclude ce-7b-17 from the test suite since OP-TEE does
not provide the necessary external command to erase TA persistent time.

Note that skipping a test will shift the numbering in the "gp" suite,
so after this patch gp_40178 will be the same as gp_40179 prior to the
patch.

Fixes: https://github.com/OP-TEE/optee_test/issues/510
Signed-off-by: Jerome Forissier <jerome@forissier.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>