xtest: Add more tests for querying signature size for C_Sign/C_SignFinal
It is possible that when querying memory that return buffer for signature
size does not have any meaningful value.

It has been observed that some PKCS#11 libraries does not initialize the
memory to zero and it is also completely fine based on specification.

Add test for the case so that it does not regress in future.

Specified in:
PKCS #11 Cryptographic Token Interface Base Specification
Version 2.40 Plus Errata 01
5.2 Conventions for functions returning output in a variable-length buffer

Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-by: Ruchika Gupta <ruchika.gupta@linaro.org>
Signed-off-by: Vesa Jääskeläinen <vesa.jaaskelainen@vaisala.com>