xtest 1013: ignore out of memory errors on large TA
Regression test 1013 (concurrency) is split in two: first, a small TA
is used in sub-case 1013.1, then a larger one is loaded in 1013.2. The
first sub-case normally passes on all platforms, but the second one may
fail on memory constrained platforms or configurations. To address this
issue, a couple of compile-time checks have been introduced (on
CFG_PAGED_USER_TA and CFG_VIRTUALIZATION). There are two problems with
this approach:
1. It is testing for features rather than actual memory constraints,
which is fragile.
2. CFG_VIRTUALIZATION isn't even exported into the TA dev kit by the
OP-TEE OS build environment, so it may happen that its value is different
in optee_os and in optee_test. Moreover, virtualization should be
transparent to TAs.

This commit removes the compile-time tests and modifies 1013.2 to ignore
any out-of-memory error when loading the TA. Other errors are still
detected.

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>