xtest: make regression_1037 compatible with OpenSSL 1.0.2o
Currently, there is a mistmatch in the OpenSSL version between OPTEE
buildroot environment and the embedded library in optee-test.
The OpenSSL version in buildroot is 1.1.1l and the embedded version is
1.0.2o.

The recent addition of regression_1037 test adds a call to
RSA_set0_key() function that is only available starting from 1.1.1
OpenSSL version.

This causes regression_1037 to fail to compile with the embedded version
of OpenSSL (WITH_OPENSSL=y and unset OPTEE_OPENSSL_EXPORT):
| tmp/work/armv8a-poky-linux/optee-test/upstream.imx-r0/git/host/xtest/regression_1000.c: In function 'set_att_key':
| tmp/work/armv8a-poky-linux/optee-test/upstream.imx-r0/git/host/xtest/regression_1000.c:2760:14: error: implicit declaration of function 'RSA_set0_key'; did you mean 'RSA_check_key'? [-Werror=implicit-function-declaration]
|  2760 |         st = RSA_set0_key(att_key, bn_n, bn_e, BN_new());
|       |              ^~~~~~~~~~~~
|       |              RSA_check_key

To fix this issue, implement RSA_set0_key() function along the test when
the OpenSSL library version is prior to 1.1.0. The implemented function
is suggested in the OpenSSL Wiki project [1]

Link: [1] https://wiki.openssl.org/index.php/OpenSSL_1.1.0_Changes
Fixes: 0915b2243 ("xtest: add regression_1037 to test remote attestation PTA")
Signed-off-by: Clement Faure <clement.faure@nxp.com>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>