regression_6000.c: detect available storage IDs at runtime
The current code relies on compile-time defines (CFG_REE_FS and
CFG_RPMB_FS) to select the storage IDs that should be used when running
the secure storage tests (TEE_STORAGE_PRIVATE, TEE_STORAGE_PRIVATE_REE,
TEE_STORAGE_PRIVATE_RPMB). This doesn't work well when xtest is run with
a TEE core that was not built with the same defines. Although that is
not a generally supported thing, the storage differences can be handled
by adding a runtime check.

Some platforms don't have access to an RPMB while others may use only
RPMB (and disable the REE storage). By adding runtime detection, and
assuming other features are the same on all platforms, we can use the
same xtest binary in a common root FS image. One exmaple is the
Linaro LEDGE Reference Platform: the root FS [1] is shared while the
firmware [2] is per-platform.

Link: [1] https://github.com/Linaro/ledge-oe-manifest
Link: [2] https://git.codelinaro.org/linaro/dependable-boot/meta-ts
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>