xtest: pkcs11: update SO PIN automatically if needed
If the system under test is using the now deprecated binary SO PIN
('0x00', '0x01', '0x02' etc), change it to the new one automatically.
This is a compatibility patch to avoid failing any test on older
systems.

Tested on QEMUv8 by running "xtest -t pkcs11" once with optee_test at
commit c0a61722df36 ("regression 1013: lower number of loops when pager
is constrained"), in other words: before the SO PIN was changed, then a
second time with a build issued from this commit.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>